shared/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lunchfork-shared"
version = "0.1.0"
description = "Shared library for lunchfork — messages, comm, geo, models"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "paho-mqtt>=1.6",
    "structlog>=23.0",
    "numpy>=1.24",
    "aiohttp>=3.9",
    "websockets>=12.0",
]

[project.optional-dependencies]
geo = [
    "rasterio>=1.3",
    "pyproj>=3.6",
    "requests>=2.31",
]
models = [
    "onnxruntime>=1.16",
]
models-cuda = [
    "onnxruntime-gpu>=1.16",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["shared*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.black]
target-version = ["py311"]
line-length = 100
